Async Updates (#184)
* Reduce line len in plugin_instance.py

* Add async to Macro & Files API

* Add async to add_bonds & add_dssp

* Fix Macro API async

* Add async to compute_hbonds

* Remove unnecessary if

* Fix create_writing_stream async

* Fix is_async check when registering callback

* Make Files API callbacks optional for async

* Fix async on files get

* Asyncify FileExplorer test plugin

* Update files.get async

+++ nanome/_internal/_process/_bonding.py
@@ -5,12 +5,24 @@ from nanome._internal._structure._io import _pdb, _sdf
 import tempfile
 import os
 
+try:
+    import asyncio
+except ImportError:
+    asyncio = False
 
 class _Bonding():
-    def __init__(self, complex_list, callback, fast_mode=None):
+    def __init__(self, plugin, complex_list, callback=None, fast_mode=None):
         self.__complexes = complex_list
         self.__framed_complexes = [complex.convert_to_frames() for complex in complex_list]
         self.__callback = callback
+        self.__future = None
+        self.__input = None
+        self.__output = None
+
+        if asyncio and plugin.is_async:
+            loop = asyncio.get_event_loop()
+            future = loop.create_future()
+            self.__future = future
 
         atom_count = 0
         if fast_mode == None:
@@ -22,7 +34,7 @@ class _Bonding():
 
     def _start(self):
         if len(self.__complexes) == 0:
-            self.__callback(self.__complexes)
+            self.__done()
             return
 
         self.__complex_idx = 0
@@ -40,6 +52,7 @@ class _Bonding():
             self.__proc.args.append('-f')
 
         self.__next()
+        return self.__future
 
     def __next(self):
         # Go to next molecule
@@ -74,7 +87,7 @@ class _Bonding():
     def __bonding_done(self, result_code):
         if result_code == -1:
             Logs.error("Couldn't execute openbabel to generate bonds. Is it installed?")
-            self.__callback(self.__complexes)
+            self.__done()
             return
         with open(self.__output.name) as f:
             lines = f.readlines()
@@ -136,8 +149,16 @@ class _Bonding():
                     new_bond._in_conformer[self.__molecule_idx] = True
 
     def __done(self):
-        self.__input.close()
-        self.__output.close()
-        os.remove(self.__input.name)
-        os.remove(self.__output.name)
-        self.__callback(self.__complexes)
+        if self.__input is not None:
+            self.__input.close()
+            os.remove(self.__input.name)
+
+        if self.__output is not None:
+            self.__output.close()
+            os.remove(self.__output.name)
+
+        if self.__callback is not None:
+            self.__callback(self.__complexes)
+
+        if self.__future is not None:
+            self.__future.set_result(self.__complexes)
